% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_binder.R
\name{AM_binder}
\alias{AM_binder}
\title{Run the binder algorithm using R (TBD)

TBD TODO : need the C version of that.}
\usage{
AM_binder(fit, weight = 0.5, with_coclustering_probability = FALSE)
}
\arguments{
\item{fit}{Output from MCMC_fit}

\item{weight}{Weight between bad and good pairs, default is 0.5.}

\item{with_coclustering_probability}{By default this function only return the index of the closest guess.
When with_coclustering_probability, the function also return the coclustering probability
matrix.}
}
\description{
Run the binder algorithm using R (TBD)

TBD TODO : need the C version of that.
}
