% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_clustering_estimation_average}
\alias{AM_clustering_estimation_average}
\title{Return maximum likelihood estimation (average)

Given a MCMC output, this function return maximum likelihood estimation.}
\usage{
AM_clustering_estimation_average(fit)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}
}
\value{
maximum likelihood estimation (average)
}
\description{
Return maximum likelihood estimation (average)

Given a MCMC output, this function return maximum likelihood estimation.
}
