% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_clustering_estimation_laugreen}
\alias{AM_clustering_estimation_laugreen}
\title{Return maximum likelihood estimation (laugreen)

Given a MCMC output, this function return maximum likelihood estimation.}
\usage{
AM_clustering_estimation_laugreen(fit, C = NULL)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}

\item{C}{used to speed up the function, can be the coclustering as returned from the command \code{\link{AM_coclustering}}.}
}
\value{
maximum likelihood estimation (laugreen)
}
\description{
Return maximum likelihood estimation (laugreen)

Given a MCMC output, this function return maximum likelihood estimation.
}
