% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_clustering_estimation_squared_loss}
\alias{AM_clustering_estimation_squared_loss}
\title{Return maximum likelihood estimation (squared_loss)

Given a MCMC output, this function return maximum likelihood estimation.}
\usage{
AM_clustering_estimation_squared_loss(fit)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}
}
\value{
maximum likelihood estimation (squared_loss)
}
\description{
Return maximum likelihood estimation (squared_loss)

Given a MCMC output, this function return maximum likelihood estimation.
}
