% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_coclustering.R
\name{AM_coclustering}
\alias{AM_coclustering}
\title{Return the co-clustering matrix}
\usage{
AM_coclustering(fit)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}
}
\value{
co-clustering matrix
}
\description{
Given an MCMC output produced by \code{AM_MCMC_fit}, this function returns the co-clustering matrix.
}
\examples{
ccm <- AM_coclustering(fit)
}
