% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_coclustering_slow}
\alias{AM_coclustering_slow}
\title{Return co-clustering slowly

Given a MCMC output, this function return co-clustering matrix}
\usage{
AM_coclustering_slow(fit)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}
}
\value{
co-clustering matrix
}
\description{
Return co-clustering slowly

Given a MCMC output, this function return co-clustering matrix
}
