% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AM_compute_stirling_ricor_abs}
\alias{AM_compute_stirling_ricor_abs}
\title{Compute the logarithm of the absolute value of the  generalized Sriling number of second Kind (mi pare) See charambeloides, using a recursive formula Devo vedere la formula}
\usage{
AM_compute_stirling_ricor_abs(n, gamma)
}
\arguments{
\item{n}{The sample size}

\item{gamma}{A positive real number \code{gamma}}
}
\value{
A vector of length n, reporting the values \code{C(gamma,n,k)} for \code{k=1,...,n}
}
\description{
There are no default values.
}
\examples{
dd= AM_compute_stirling_ricor_abs(11,10)
print(dd)
}
\keyword{cluster}
\keyword{number}
\keyword{of}
\keyword{prior}
