% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AM_compute_stirling_ricor_log}
\alias{AM_compute_stirling_ricor_log}
\title{Compute ...}
\usage{
AM_compute_stirling_ricor_log(n, gamma)
}
\arguments{
\item{n}{The sample size}

\item{gamma}{A positive real number \code{gamma}}
}
\value{
A vector of length n, reporting the values ... for \code{k=1,...,n}
}
\description{
There are no default values.
}
\examples{
dd= AM_compute_stirling_ricor_log(11,10)
print(dd)
}
\keyword{cluster}
\keyword{number}
\keyword{of}
\keyword{prior}
