% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_demo.R
\name{AM_demo_mvb_poi}
\alias{AM_demo_mvb_poi}
\title{Returns an example of \code{AM_mcmc_fit} output produced by the multivariate bernoulli model}
\usage{
AM_demo_mvb_poi()
}
\value{
The return value is a \code{\link{AM_mcmc_output}} object.
}
\description{
This function allows us to generate a sample output of fitting the multivariate bernoulli model. No arguments are needed to be passed.
The purpose of this function is to serve as a demo for users to understand the model's output, without diving too deep into details. By default,
this demo generates a sample dataset of dimension 500x4, where the MCMC sampler is specified to run for 2000 iterations, with a burn-in of 1000, and a thinning interval of 10. All possible outputs
that can be produced by \code{\link{AM_mcmc_fit}} are returned.
}
\examples{
mvb_output <- AM_demo_mvb_poi()
}
\keyword{demo}
