% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_demo.R
\name{AM_demo_uvn_poi}
\alias{AM_demo_uvn_poi}
\title{AM_demo_uvn_poi

Returns AM_MCMC_OUTPUT Object from a running demo of Univariate Poisson Gamma.}
\usage{
AM_demo_uvn_poi(
  n = 500,
  mcmc_params = AM_mcmc_parameters(niter = 2000, burnin = 1000, thin = 10, verbose = 0,
    output = c("ALL"))
)
}
\arguments{
\item{n}{Number of sample}

\item{mcmc_params}{AM_mcmc_parameters object to fit with}
}
\description{
AM_demo_uvn_poi

Returns AM_MCMC_OUTPUT Object from a running demo of Univariate Poisson Gamma.
}
\keyword{demo}
