% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_extract.R
\name{AM_extract}
\alias{AM_extract}
\title{Extract values within a AM_mcmc_output object.}
\usage{
AM_extract(object, targets, iterations = NULL, debug = FALSE)
}
\arguments{
\item{object}{a \code{\link{AM_mcmc_output}} object}

\item{targets}{List of variables to extract (ie. K, M, Mna, Tau_mu).}

\item{iterations}{Can specify particular iterations to extracts, NULL for all.}

\item{debug}{Activate log to}
}
\description{
Due to the complexity of AntMAN outputs, AM_mcmc_output object can be difficult
to handle. The AM_extract function ease access of particular variables within
AM_mcmc_output object.
}
