% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{AM_mcmc_configuration}
\alias{AM_mcmc_configuration}
\title{S3 class AM_mcmc_configuration}
\value{
\code{\link{AM_mcmc_configuration}}
}
\description{
Output type of return values from  \code{\link{AM_mcmc_parameters}}.
}
\seealso{
\code{\link{AM_mcmc_fit}}
}
