% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_mcmc_output}
\alias{AM_mcmc_output}
\title{S3 class AM_mcmc_output.}
\description{
Output type of return values from  \code{\link{AM_mcmc_fit}}. See paper for the moment.
}
\seealso{
\code{\link{AM_mcmc_fit}}
}
