% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_components_prior.R
\name{AM_mix_components_prior}
\alias{AM_mix_components_prior}
\title{S3 class AM_mix_components_prior}
\value{
\code{\link{AM_mix_components_prior}}
}
\description{
Object returned by \code{AM_mix_components_prior_*}.
}
\seealso{
\code{\link{AM_mix_components_prior_dirac}},
\code{\link{AM_mix_components_prior_negbin}},
\code{\link{AM_mix_components_prior_pois}}
}
