% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_components_prior.R
\name{AM_mix_components_prior_dirac}
\alias{AM_mix_components_prior_dirac}
\title{Generate a configuration object that contains a Point mass prior}
\usage{
AM_mix_components_prior_dirac(Mstar)
}
\arguments{
\item{Mstar}{Fixed value  \eqn{M^*} for the number of components.}
}
\value{
An \code{\link{AM_mix_components_prior}} object. This is a configuration list to be used as \code{mix_components_prior} argument for \code{\link{AM_mcmc_fit}}.
}
\description{
Generate a configuration object that assigns a Point mass prior to the number of mixture components.
This is the simplest option and it requires users to specify a value \eqn{M^*}
such that \eqn{Pr(M=M^* =1}.
}
\examples{

AM_mix_components_prior_dirac (Mstar=3)
}
\seealso{
\code{\link{AM_mcmc_fit}}
}
\keyword{prior}
