% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_hyperparams.R
\name{AM_mix_hyperparams}
\alias{AM_mix_hyperparams}
\title{S3 class AM_mix_hyperparams}
\value{
\code{\link{AM_mix_hyperparams}}
}
\description{
Object type returned by \code{AM_mix_hyperparams_*} commands.
}
\seealso{
\code{\link{AM_mix_hyperparams_unipois}}, \code{\link{AM_mix_hyperparams_uninorm}}, \code{\link{AM_mix_hyperparams_multiber}},
\code{\link{AM_mix_hyperparams_multinorm}}
}
