% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_hyperparams.R
\name{AM_mix_hyperparams_multiber}
\alias{AM_mix_hyperparams_multiber}
\title{multivariate Bernoulli mixture hyperparameters (Latent Class Analysis)}
\usage{
AM_mix_hyperparams_multiber(a0, b0)
}
\arguments{
\item{a0}{The a0 hyperparameters.}

\item{b0}{The b0 hyperparameters.}
}
\value{
An \code{\link{AM_mix_hyperparams}} object. This is a configuration list to be used as \code{mix_kernel_hyperparams} argument for \code{\link{AM_mcmc_fit}}.
}
\description{
Generate a configuration object that defines the prior hyperparameters for a mixture of multivariate Bernoulli.
If the dimension of the data is P, then the prior is a product of P independent Beta distributions, Beta(\eqn{a_{0i},b_{0i}}). Therefore,
the vectors of hyperparameters, a0 and b0,  are P-dimensional. Default is (a0= c(1,....,1),b0= c(1,....,1)).
}
\examples{
AM_mix_hyperparams_multiber (a0= c(1,1,1,1),b0= c(1,1,1,1))
}
