% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_weights_prior.R
\name{AM_mix_weights_prior}
\alias{AM_mix_weights_prior}
\title{S3 class AM_mix_weights_prior}
\value{
\code{\link{AM_mix_weights_prior}}
}
\description{
Object type returned by \code{AM_mix_weights_prior_*} commands.
}
\seealso{
\code{\link{AM_mix_weights_prior_gamma}}
}
