% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_multinorm_mix_hyperparams}
\alias{AM_multinorm_mix_hyperparams}
\title{Multivariate Normal Mixture Hyperparameters.}
\usage{
AM_multinorm_mix_hyperparams(mu0 = NULL, ka0 = NULL, nu0 = NULL,
  Lam0 = NULL)
}
\arguments{
\item{mu0}{The hyperparameter \eqn{\boldsymbol m_0}.}

\item{ka0}{The hyperparameter \eqn{\kappa_0}.}

\item{nu0}{The hyperparameter \eqn{\nu_0}.}

\item{Lam0}{The hyperparameter \eqn{\Lambda_0}.}
}
\value{
A list to be used as \code{mix_kernel_hyperparams} argument for \code{mcmc_fit}.
}
\description{
Speicifies the hyperparameter for the conjugate prior for a mixture of Multivariate Normals. I NEED TO NOW THE PARAMETRIZATION
\deqn{\pi(\boldsymbol \mu, \boldsymbol \Sigma\mid\boldsymbol m_0,\kappa_0,\nu_0,\boldsymbol \Lambda_0)= \pi_{\mu}(\boldsymbol \mu|\boldsymbol \Sigma,\boldsymbol m_0,\kappa_0)\pi_{\Sigma}(\boldsymbol \Sigma \mid \nu_0,\boldsymbol \Lambda_0)}
with \code{mu0} is \eqn{\boldsymbol m_0}, \code{ka0} is \eqn{\kappa_0},
\code{nu0} is \eqn{\nu_0}, \code{Lam0} is \eqn{\Lambda_0}.
}
\examples{
AM_multinorm_mix_hyperparams ()
}
