% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_density}
\alias{AM_plot_density}
\title{plot the density of variables from AM_mcmc_output object.}
\usage{
AM_plot_density(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{An AM_mcmc_output fit object, produced by calling \code{AM_MCMC_fit}.}

\item{tags}{A list of variables to consider. This function only produces meaningful plots for variables that have fixed dimension across the draws. If not specified, plots pertaining to M and K will be produced.}

\item{title}{Title for the plot.}
}
\value{
a ggplot object visualising the posterior density of variables M and K.
}
\description{
Given an object of class fit, AM_plot_density plots the posterior density of variables M and K combined across all chains. AM_plot_density makes use
of bayesplot's plotting function mcmc_areas.
}
