% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_mvb_cluster_frequency}
\alias{AM_plot_mvb_cluster_frequency}
\title{Visualise the cluster frequency plot for the multivariate bernoulli model}
\usage{
AM_plot_mvb_cluster_frequency(
  fit,
  y,
  x_lim_param = c(0.8, 7.2),
  y_lim_param = c(0, 1)
)
}
\arguments{
\item{fit}{An \code{\link{AM_mcmc_output}} fit object, produced by calling \code{AM_mcmc_fit}.}

\item{y}{A matrix containing the y observations which produced fit.}

\item{x_lim_param}{A vector with two elements describing the plot's x_axis scale, e.g. c(0.8, 7.2).}

\item{y_lim_param}{A vector with two elements describing the plot's y_axis scale, e.g. c(0, 1).}
}
\value{
No return value. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, and the data the model was fit on, this function will produce a cluster frequency plot for the multivariate bernoulli model.
}
