% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_pmf}
\alias{AM_plot_pmf}
\title{Plot the probability mass function of variables from \code{\link{AM_mcmc_output}} object}
\usage{
AM_plot_pmf(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{An \code{\link{AM_mcmc_output}} object, produced by calling \code{\link{AM_mcmc_fit}}.}

\item{tags}{A list of variables to consider. If not specified, the pmf of both M and K will be plotted.}

\item{title}{Title for the plot.}
}
\value{
No return value. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, AM_plot_pmf plots the posterior probability mass function of the specified variables.
}
