% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_plot_similarity_matrix}
\alias{AM_plot_similarity_matrix}
\title{Plot the Similarity Matrix

Given a MCMC output, this function will produce an image of the Similarity Matrix}
\usage{
AM_plot_similarity_matrix(C, ...)
}
\arguments{
\item{fit}{a \code{\link{AM_mcmc_output}} object}
}
\description{
Plot the Similarity Matrix

Given a MCMC output, this function will produce an image of the Similarity Matrix
}
