% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_traces}
\alias{AM_plot_traces}
\title{plot traces of variables from AM_mcmc_output object}
\usage{
AM_plot_traces(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{a AM_mcmc_output object}

\item{tags}{A list of variables to consider}

\item{title}{Title for the plot}
}
\value{
TBD
}
\description{
TBD
}
