% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_prior.R
\name{AM_prior}
\alias{AM_prior}
\title{S3 class AM_prior}
\value{
\code{\link{AM_prior}}
}
\description{
Object type returned by \code{AM_prior_*} commands.
}
\seealso{
\code{\link{AM_prior_K_Delta}}, \code{\link{AM_prior_K_Pois}},  \code{\link{AM_prior_K_NegBin}}
}
