% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_helpers.R
\name{AM_sample_multibin}
\alias{AM_sample_multibin}
\title{AM_sample_multibin}
\usage{
AM_sample_multibin(n, d, pro, TH)

AM_sample_multibin(n, d, pro, TH)
}
\description{
AM_sample_multibin

AM_sample_multibin
}
\keyword{internal}
