% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_sample.R
\name{AM_sample_unipois}
\alias{AM_sample_unipois}
\title{AM_sample_unipois}
\usage{
AM_sample_unipois(n = 1000, pro = c(0.2, 0.5, 0.3), mth = c(5, 25, 50))
}
\description{
AM_sample_unipois
}
\keyword{internal}
