% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_unibin_mix_hyperparams}
\alias{AM_unibin_mix_hyperparams}
\title{Univariate Binomial Mixture Hyperparameters.

Generate a configuration object that define Univariate Binomial Mixture Hyperparameters such as :
\deqn{ \pi(\theta\mid \alpha,\beta)=\frac{\Gamma(\alpha+\beta)}{\Gamma(\alpha)\Gamma(\beta)} \tau^{\alpha-1}\left( 1-\tau \right)^{\beta-1} , \qquad 0\le\tau\le1}
with \eqn{\alpha} is \code{a0},\eqn{\beta} is \code{b0}, and N is ... oups I need to add more infos in those sections.
a0=1 b0=1}
\usage{
AM_unibin_mix_hyperparams(a0, b0, N)
}
\arguments{
\item{a0}{The a0 hyperparameter.}

\item{b0}{The b0 hyperparameter.}

\item{N}{size of binomial .}
}
\value{
A list to be used as \code{mix_kernel_hyperparams} argument for \code{mcmc_fit}.
}
\description{
Univariate Binomial Mixture Hyperparameters.

Generate a configuration object that define Univariate Binomial Mixture Hyperparameters such as :
\deqn{ \pi(\theta\mid \alpha,\beta)=\frac{\Gamma(\alpha+\beta)}{\Gamma(\alpha)\Gamma(\beta)} \tau^{\alpha-1}\left( 1-\tau \right)^{\beta-1} , \qquad 0\le\tau\le1}
with \eqn{\alpha} is \code{a0},\eqn{\beta} is \code{b0}, and N is ... oups I need to add more infos in those sections.
a0=1 b0=1
}
\examples{
AM_unibin_mix_hyperparams (a0=1,b0=1,N=100)
}
