% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_unipois_mix_hyperparams}
\alias{AM_unipois_mix_hyperparams}
\title{Univariate Poisson Mixture Hyperparameters.}
\usage{
AM_unipois_mix_hyperparams(alpha0, beta0)
}
\arguments{
\item{alpha0}{The shape  hyperparameter \eqn{\alpha_0}.}

\item{beta0}{The  rate hyperparameter \eqn{\beta_0}.}
}
\value{
A list to be used as \code{mix_kernel_hyperparams} argument for \code{mcmc_fit}.
}
\description{
Univariate Poisson Mixture Hyperparameters.
}
\details{
By default alpha0=1 and beta0=1.
}
\examples{
AM_unipois_mix_hyperparams (alpha0=2, beta0=0.2)
}
