\name{AntMAN-package}
\alias{AntMAN-package}
\alias{AntMAN}
\docType{package}
\title{
\packageTitle{AntMAN}
}
\description{
\packageDescription{AntMAN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AntMAN}
\packageIndices{AntMAN}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AntMAN}

Maintainer: \packageMaintainer{AntMAN}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
print("Example");
}
