% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{package}
\name{AntMAN}
\alias{AntMAN}
\alias{AntMAN-package}
\title{AntMAN: A package for fitting Finite Bayesian Mixture model with random number of component.}
\description{
AntMAN: Anthology of Mixture ANalysis tools 
AntMan is a R package to fit Finite Bayesian Mixture  model with random number of component. 
It is an alternative to Reversible Jump. 
Different kernels: Univariate Gaussian, Univariate Poisson, Univariate Binomial, Multivariate Gaussian, 
Multivariate Bernoulli (Latent Class Analysis). Conjugate priors.  
 We allow for different choices for the prior on the number of components: 
Shifted Poisson, Negative Binomial, and Point Masses (i.e. mixtures with fixed number of components).
}
\section{Prior functions}{

The Prior functions ...
}

