\name{brain}
\alias{brain}
\docType{data}
\title{Teen Brain Images from the National Institutes of Health, U.S.}
\description{
	Picture of brain activity.
}
\usage{data(brain)}
\format{
  A list of pixel taken from a YxY image in RGB format.
}
\source{
  https://www.flickr.com/photos/nida-nih/29741916012
}
\references{
    Crowley TJ, Dalwani MS, Mikulich-Gilbertson SK, Young SE, Sakai JT, Raymond KM, et al. (2015) Adolescents' Neural Processing of Risky Decisions: Effects of Sex and Behavioral Disinhibition. PLoS ONE 10(7): e0132322. doi:10.1371/journal.pone.0132322
}
\examples{
data(brain)
}
\keyword{datasets}
