% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{data}
\name{carcinoma}
\alias{carcinoma}
\title{carcinoma

carcinoma is ....}
\format{A data frame with X rows and Y variables.}
\source{
URL?
}
\usage{
carcinoma
}
\description{
carcinoma

carcinoma is ....
}
\examples{
 data(carcinoma)
 
}
\keyword{datasets}
