% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{galaxy
 
carcinoma is .... 
 
 @format A data frame with X rows and Y variables.
 
 @source URL?
 
 @examples
 data(carcinoma)
 
 @keyword datasets}
\format{An object of class \code{numeric} of length 82.}
\usage{
galaxy
}
\description{
galaxy
 
carcinoma is .... 
 
 @format A data frame with X rows and Y variables.
 
 @source URL?
 
 @examples
 data(carcinoma)
 
 @keyword datasets
}
\keyword{datasets}
