% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL_utilities.R
\name{IAM_compute_stirling_ricor_abs}
\alias{IAM_compute_stirling_ricor_abs}
\title{Compute the logarithm of the absolute value of the  generalized Sriling number of second Kind (mi pare) See charambeloides, using a recursive formula Devo vedere la formula}
\usage{
IAM_compute_stirling_ricor_abs(n, gamma)
}
\arguments{
\item{n}{The sample size}

\item{gamma}{A positive real number \code{gamma}}
}
\value{
A vector of length n, reporting the values \code{C(gamma,n,k)} for \code{k=1,...,n}
}
\description{
There are no default values.
}
\keyword{internal}
