% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL_utilities.R
\name{IAM_compute_stirling_ricor_log}
\alias{IAM_compute_stirling_ricor_log}
\title{Compute stirling ricor log}
\usage{
IAM_compute_stirling_ricor_log(n, gamma)
}
\arguments{
\item{n}{The sample size}

\item{gamma}{A positive real number \code{gamma}}
}
\value{
A vector of length n, reporting the values ... for \code{k=1,...,n}
}
\description{
There are no default values.
}
\keyword{internal}
