% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL_utilities.R
\name{IAM_mcmc_error}
\alias{IAM_mcmc_error}
\title{Internal function used to compute the MCMC Error as a batch mean.}
\usage{
IAM_mcmc_error(X)
}
\arguments{
\item{X}{is a chain}
}
\value{
the MCMC Error (sqrt(sigma2) / sqrt(N))
}
\description{
Internal function used to compute the MCMC Error as a batch mean.
}
\keyword{internal}
