% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERNAL_utilities.R
\name{IAM_mcmc_neff}
\alias{IAM_mcmc_neff}
\title{IAM_mcmc_neff MCMC Parameters}
\usage{
IAM_mcmc_neff(unichain)
}
\arguments{
\item{unichain}{}
}
\value{
Effective Sample Size
}
\description{
TBD
}
\keyword{internal}
