% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{list_values}
\alias{list_values}
\title{Internal function that produces a string from a list of values}
\usage{
list_values(x)
}
\arguments{
\item{x}{a list of values}
}
\description{
Internal function that produces a string from a list of values
}
\keyword{internal}
