% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_prior.R
\name{plot.AM_prior}
\alias{plot.AM_prior}
\title{plot AM_prior}
\usage{
\method{plot}{AM_prior}(x, ...)
}
\arguments{
\item{x}{an \code{\link{AM_prior}} object. See \code{\link{AM_prior_K_Delta}}, \code{\link{AM_prior_K_NegBin}},
\code{\link{AM_prior_K_Pois}} for more details.}

\item{...}{all additional parameters are ignored.}
}
\value{
NULL. Called for side effects.
}
\description{
plot the prior on the number of clusters for a given \code{\link{AM_prior}} object.
}
