% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{summary.AM_mcmc_configuration}
\alias{summary.AM_mcmc_configuration}
\title{summary AM_mcmc_configuration

Print infos on AM_mcmc_configuration Object}
\usage{
\method{summary}{AM_mcmc_configuration}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mcmc_configuration}} object}

\item{...}{all additionnal parameters are ignored}
}
\description{
summary AM_mcmc_configuration

Print infos on AM_mcmc_configuration Object
}
