% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{summary.AM_mcmc_output}
\alias{summary.AM_mcmc_output}
\title{summary AM_mcmc_output

Print some useful informations about the mcmc results}
\usage{
\method{summary}{AM_mcmc_output}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mcmc_output}} object}

\item{...}{all additionnal parameters are ignored}
}
\description{
summary AM_mcmc_output

Print some useful informations about the mcmc results
}
