% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_components_prior.R
\name{summary.AM_mix_components_prior}
\alias{summary.AM_mix_components_prior}
\title{summary AM_mix_components_prior

Print infos on AM_mix_components_prior Object}
\usage{
\method{summary}{AM_mix_components_prior}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mix_components_prior}} object}

\item{...}{all additional parameters are ignored}
}
\description{
summary AM_mix_components_prior

Print infos on AM_mix_components_prior Object
}
