% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_hyperparams.R
\name{summary.AM_mix_hyperparams}
\alias{summary.AM_mix_hyperparams}
\title{summary AM_mix_hyperparams

Print infos on AM_mix_hyperparams Object}
\usage{
\method{summary}{AM_mix_hyperparams}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mix_hyperparams}} object}

\item{...}{all additionnal parameters are ignored}
}
\description{
summary AM_mix_hyperparams

Print infos on AM_mix_hyperparams Object
}
