% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_weights_prior.R
\name{summary.AM_mix_weights_prior}
\alias{summary.AM_mix_weights_prior}
\title{summary AM_mix_weights_prior

Print infos on AM_mix_weights_prior Object}
\usage{
\method{summary}{AM_mix_weights_prior}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mix_weights_prior}} object}

\item{...}{all additionnal parameters are ignored}
}
\description{
summary AM_mix_weights_prior

Print infos on AM_mix_weights_prior Object
}
