% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AM_emp_bayes_uninorm}
\alias{AM_emp_bayes_uninorm}
\title{this function compute the hyperparameters of an Normal-Inverse-Gamma
distribution using an empirical Bayes approach.}
\usage{
AM_emp_bayes_uninorm(y, scEmu = 1, scEsig2 = 3, CVsig2 = 3)
}
\arguments{
\item{y}{The data y}

\item{scEmu}{a positive value scEmu (default =1) such that marginally E(mu)=(sample variance)*scEmu}

\item{scEsig2}{a positive value scEsig2 (default=3) such that marginally E(sig2)=(sample variance)*scEsig2}

\item{CVsig2}{The coefficient of variation of sig2 (default =3), CVsig2}
}
\value{
Parameters of the normal inverse Gamma prior according to the empirical Bayes approach.
}
\description{
this function compute the hyperparameters of an Normal-Inverse-Gamma
distribution using an empirical Bayes approach.
}
