% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_chaincor}
\alias{AM_plot_chaincor}
\title{Plot the Autocorrelation

Given a MCMC output, this function will produce acf bar}
\usage{
AM_plot_chaincor(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{a AM_mcmc_output object}

\item{tags}{A list of variables to consider}

\item{title}{Title for the plot}
}
\description{
Plot the Autocorrelation

Given a MCMC output, this function will produce acf bar
}
