% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_density}
\alias{AM_plot_density}
\title{Plot the density of variables from \code{\link{AM_mcmc_output}} object}
\usage{
AM_plot_density(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{An \code{\link{AM_mcmc_output}} fit object, produced by calling \code{\link{AM_mcmc_fit}}.}

\item{tags}{A list of variables to consider. This function only produces meaningful plots for variables that have fixed dimension across the draws.}

\item{title}{Title for the plot.}
}
\value{
a ggplot object visualising the posterior density of the specified variables.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, AM_plot_density plots the posterior density of the specified variables of interest. AM_plot_density makes use
of bayesplot's plotting function mcmc_areas \insertCite{bayesplot}{AntMAN}.
}
