% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{AM_plot_similarity_matrix}
\alias{AM_plot_similarity_matrix}
\title{Plot the Similarity Matrix

Given a MCMC output, this function will produce an image of the Similarity Matrix}
\usage{
AM_plot_similarity_matrix(C, ...)
}
\arguments{
\item{C}{coclustering as returned from the command \code{\link{AM_coclustering}}.}

\item{...}{All additional parameters wil lbe pass to the image command.}
}
\description{
Plot the Similarity Matrix

Given a MCMC output, this function will produce an image of the Similarity Matrix
}
