% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_similarity_matrix}
\alias{AM_plot_similarity_matrix}
\title{Plot the Similarity Matrix}
\usage{
AM_plot_similarity_matrix(x, loss, ...)
}
\arguments{
\item{x}{An \code{\link{AM_mcmc_output}} fit object, produced by calling \code{AM_mcmc_fit}.}

\item{loss}{Loss function to minimise. Specify either "VI" or "binder". If not specified, the default loss
to minimise is "binder".}

\item{...}{All additional parameters wil lbe pass to the image command.}
}
\value{
No return value. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, this function will produce an image of the Similarity Matrix.
}
