% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_plot.R
\name{AM_plot_traces}
\alias{AM_plot_traces}
\title{Plot traces of variables from an \code{\link{AM_mcmc_output}} object}
\usage{
AM_plot_traces(x, tags = NULL, title = "MCMC Results")
}
\arguments{
\item{x}{An \code{\link{AM_mcmc_output}} fit object, produced by calling \code{\link{AM_mcmc_fit}}.}

\item{tags}{A list of variables to consider. This function only produces meaningful plots for variables that have fixed dimension across the draws. If not specified, plots pertaining to M and K will be produced.}

\item{title}{Title for the plot}
}
\value{
No return value. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, \code{\link{AM_plot_traces}} visualises the traceplots of the specified variables involved in the MCMC inference.
AM_plot_traces is built upon bayesplot's mcmc_trace \insertCite{bayesplot}{AntMAN}.
}
