\name{AntMan-package}
\alias{AntMan-package}
\alias{AntMan}
\docType{package}
\title{
\packageTitle{AntMan}
}
\description{
\packageDescription{AntMan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AntMan}
\packageIndices{AntMan}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AntMan}

Maintainer: \packageMaintainer{AntMan}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
print("Example");
}
