% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{data}
\name{brain}
\alias{brain}
\title{brain
 
Teen Brain Images from the National Institutes of Health, U.S.
 
Picture of brain activity from a teenager consumming drugs. 
 
 @format  A list of pixel taken from a YxY image in RGB format.
 
 @source https://www.flickr.com/photos/nida-nih/29741916012
 @references Crowley TJ, Dalwani MS, Mikulich-Gilbertson SK, Young SE, Sakai JT, Raymond KM, et al. (2015) Adolescents' Neural Processing of Risky Decisions: Effects of Sex and Behavioral Disinhibition. PLoS ONE 10(7): e0132322. doi:10.1371/journal.pone.0132322
 
 @examples
 data(brain)
 
 @keyword datasets}
\format{An object of class \code{list} of length 2.}
\usage{
brain
}
\description{
brain
 
Teen Brain Images from the National Institutes of Health, U.S.
 
Picture of brain activity from a teenager consumming drugs. 
 
 @format  A list of pixel taken from a YxY image in RGB format.
 
 @source https://www.flickr.com/photos/nida-nih/29741916012
 @references Crowley TJ, Dalwani MS, Mikulich-Gilbertson SK, Young SE, Sakai JT, Raymond KM, et al. (2015) Adolescents' Neural Processing of Risky Decisions: Effects of Sex and Behavioral Disinhibition. PLoS ONE 10(7): e0132322. doi:10.1371/journal.pone.0132322
 
 @examples
 data(brain)
 
 @keyword datasets
}
\keyword{datasets}
