% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntMANDatasets.R
\docType{data}
\name{brain}
\alias{brain}
\title{Teen Brain Images from the National Institutes of Health, U.S.}
\format{
A list that contains \code{dim} a (W:width,H:height) pair, and  \code{pic} a data frame (W*H pixels image in RGB format).
}
\source{
https://www.flickr.com/photos/nida-nih/29741916012
}
\usage{
brain
}
\description{
Picture of brain activities from a teenager consuming drugs.
}
\examples{
 data(brain)
 
}
\references{
Crowley TJ, Dalwani MS, Mikulich-Gilbertson SK, Young SE, Sakai JT, Raymond KM, et al. (2015) Adolescents' Neural Processing of Risky Decisions: Effects of Sex and Behavioral Disinhibition. PLoS ONE 10(7): e0132322. doi:10.1371/journal.pone.0132322
}
\keyword{datasets}
