% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{data}
\name{carcinoma}
\alias{carcinoma}
\title{carcinoma
 
carcinoma is .... 
 
 @format A data frame with X rows and Y variables.
 
 @source URL?
 
 @examples
 data(carcinoma)
 
 @keyword datasets}
\format{An object of class \code{matrix} with 118 rows and 7 columns.}
\usage{
carcinoma
}
\description{
carcinoma
 
carcinoma is .... 
 
 @format A data frame with X rows and Y variables.
 
 @source URL?
 
 @examples
 data(carcinoma)
 
 @keyword datasets
}
\keyword{datasets}
