% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntMANDatasets.R
\docType{data}
\name{carcinoma}
\alias{carcinoma}
\title{Carcinoma dataset}
\format{
A data frame with 118 rows and 7 variables (from A to G).
}
\usage{
carcinoma
}
\description{
The carcinoma data from Agresti (2002, 542) consist of seven dichotomous variables representing
the ratings by seven pathologists of 118 slides on the presence or absence of carcinoma.
The purpose of studying this data is to model "interobserver agreement" by examining how
subjects might be divided into groups depending upon the consistency of their diagnoses.
}
\examples{
 data(carcinoma)
 
}
\references{
Agresti A (2002). Categorical Data Analysis. John Wiley & Sons, Hoboken.
}
\keyword{datasets}
