% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy velocities dataset}
\format{A data frame with X rows and Y variables.}
\source{
Roeder, K. (1990) Density estimation with confidence sets exemplified by superclusters and voids in the galaxies, Journal of the American Statistical Association, 85: 617-624.
}
\usage{
galaxy
}
\description{
This data set consider physical information on velocities (km/second) for
82 galaxies reported by Roeder (1990). These are drawn from six well-separated
conic sections of the Corona Borealis region.
}
\examples{
 data(galaxy)
}
\keyword{datasets}
