% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntMANDatasets.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy velocities dataset}
\format{
A data frame with X rows and Y variables.

A numeric vector giving the speed of galaxies (1000*(km/second))
}
\source{
Roeder, K. (1990). Density estimation with confidence sets exemplified by superclusters and voids in the galaxies, Journal of the American Statistical Association, 85: 617-624.
}
\usage{
galaxy
}
\description{
This data set considers the physical information of velocities (10^3 km/second) for
82 galaxies reported by Roeder (1990). These are drawn from six well-separated
conic sections of the Corona Borealis region.
}
\examples{
 data(galaxy)
}
\keyword{datasets}
