% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntManAPI.R
\name{plot.AM_mcmc_output}
\alias{plot.AM_mcmc_output}
\title{plot AM_mcmc_output}
\usage{
\method{plot}{AM_mcmc_output}(x, ...)
}
\arguments{
\item{x}{a AM_mcmc_output object}

\item{...}{all additionnal parameters passed to image command.}
}
\description{
plot some useful informations about the mcmc results
}
