% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{plot.AM_mcmc_output}
\alias{plot.AM_mcmc_output}
\title{plot AM_mcmc_output}
\usage{
\method{plot}{AM_mcmc_output}(x, ...)
}
\arguments{
\item{x}{an \code{\link{AM_mcmc_output}} object.}

\item{...}{all additional parameters are ignored.}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, this function plots some useful information about the MCMC results
regarding \eqn{M} and \eqn{K}. Besides the PMFs, some of the diagnostic plots of the MCMC chain are visualised.
}
