% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntMANDatasets.R
\docType{data}
\name{said}
\alias{said}
\title{Usage frequency of the word "said" in the Brown corpus}
\format{
A list with 500 observations on the frequency of said in different texts.
}
\source{
https://www.kaggle.com/nltkdata/brown-corpus
}
\usage{
said
}
\description{
Usage frequency of the word "said" in the Brown corpus
}
\examples{
 data(said)
}
\references{
Francis, W., and Kucera, H. (1982) Frequency Analysis of English Usage, Houghton Mifflin Company, Boston.
}
\keyword{datasets}
