% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{summary.AM_mcmc_configuration}
\alias{summary.AM_mcmc_configuration}
\title{summary information of the AM_mcmc_configuration object}
\usage{
\method{summary}{AM_mcmc_configuration}(object, ...)
}
\arguments{
\item{object}{an \code{\link{AM_mcmc_configuration}} object.}

\item{...}{all additional parameters are ignored}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_configuration}} object, this function prints the summary information
of the specified mcmc configuration.
}
\seealso{
\code{\link{AM_mcmc_parameters}}
}
