% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mcmc.R
\name{summary.AM_mcmc_output}
\alias{summary.AM_mcmc_output}
\title{summary information of the AM_mcmc_output object}
\usage{
\method{summary}{AM_mcmc_output}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_mcmc_output}} object}

\item{...}{all additional parameters are ignored}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_mcmc_output}} object, this function prints the summary information
pertaining to the given model output.
}
\seealso{
\code{\link{AM_mcmc_fit}}, \code{\link{AM_mcmc_refit}}
}
