% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_hyperparams.R
\name{summary.AM_mix_hyperparams}
\alias{summary.AM_mix_hyperparams}
\title{summary information of the AM_mix_hyperparams object}
\usage{
\method{summary}{AM_mix_hyperparams}(object, ...)
}
\arguments{
\item{object}{an \code{\link{AM_mix_hyperparams}} object.}

\item{...}{all additional parameters are ignored.}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_mix_hyperparams}} object, this function prints the summary information
of the specified mixture hyperparameters.
}
\seealso{
\code{\link{AM_mix_hyperparams}}
}
