% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_mix_weights_prior.R
\name{summary.AM_mix_weights_prior}
\alias{summary.AM_mix_weights_prior}
\title{summary information of the AM_mix_weights_prior object}
\usage{
\method{summary}{AM_mix_weights_prior}(object, ...)
}
\arguments{
\item{object}{an \code{\link{AM_mix_weights_prior}} object.}

\item{...}{all additional parameters are ignored.}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_mix_weights_prior}} object, this function prints the summary information
of the specified mixture weights prior.
}
\seealso{
\code{\link{AM_mix_weights_prior}}
}
