% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_prior.R
\name{summary.AM_prior}
\alias{summary.AM_prior}
\title{summary AM_prior

Print infos on AM_prior Object}
\usage{
\method{summary}{AM_prior}(object, ...)
}
\arguments{
\item{object}{a \code{\link{AM_prior}} object}

\item{...}{all additionnal parameters are ignored}
}
\description{
summary AM_prior

Print infos on AM_prior Object
}
