% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_prior.R
\name{summary.AM_prior}
\alias{summary.AM_prior}
\title{summary information of the AM_prior object}
\usage{
\method{summary}{AM_prior}(object, ...)
}
\arguments{
\item{object}{an \code{\link{AM_prior}} object. See \code{\link{AM_prior_K_Delta}}, \code{\link{AM_prior_K_NegBin}},
\code{\link{AM_prior_K_Pois}} for more details.}

\item{...}{all additional parameters are ignored.}
}
\value{
NULL. Called for side effects.
}
\description{
Given an \code{\link{AM_prior}} object, this function prints the summary information of the specified prior on the number of clusters.
}
\seealso{
\code{\link{AM_prior}}
}
